+++ BSP/Inc/UART.h
@@ -16,6 +16,25 @@
 
 #include "stm32l4xx_hal.h" // stm32l4xx HAL library, substituing for the agnostic approach
 
+
+/* Function Descriptions:
+ * Init initializes the UART peripheral given the configuration
+ * DeInit deinitializes the UART peripheral meaning the configuration is reset
+ * Start starts the UART peripheral meaning it is able to interrupt and transmit/receive
+ * Stop stops the UART peripheral meaning the configuration is unchanged but inactive
+ * Send queues data to be transmitted when the UART peripheral is available
+ * Recv reads received data from the RX queue
+ *
+ * How to Use This Driver:
+ * 1. Init to initialize the UART peripheral
+ * 2. Start to activate the UART peripheral
+ * 3. Send/Recv operations
+ * 4. Stop/Start as needed
+ * 5. DeInit to deinitialize
+ * 6. Init to start over...
+ */
+
+
 // Return codes for UART operations
 typedef enum {
     UART_ERR,
@@ -25,40 +44,22 @@ typedef enum {
     UART_EMPTY,
 } uart_status_t;
 
+
 // UART handlers - similar to how CAN exposes its handlers
 #ifdef UART4
 extern UART_HandleTypeDef* huart4;
 #endif /* UART4 */
 
+#ifdef UART5
+extern UART_HandleTypeDef* huart5;
+#endif /* UART5 */
+
 // Function prototypes with consistent parameters and return types
-/**
- * @brief Initializes UART peripheral and internal TX queue.
- * Hardware configuration must be performed outside of this function.
- *
- * @param handle pointer to the UART handle
- * @param rxQueue pointer to the receive queue
- * @return uart_status_t
- */
 uart_status_t uart_init(UART_HandleTypeDef* handle, QueueHandle_t* rxQueue);
-
-/**
- * @brief Transmits data over UART. If transmission is in progress,
- * data will be queued in internal TX queue.
- * @param handle pointer to the UART handle
- * @param data pointer to the data buffer that will be written/transmitted
- * @param length of the buffer that will be written/transmitted
- * @param blocking whether the operation should be blocking
- * @return uart_status_t
- */
+// uart_status_t uart_deinit(UART_HandleTypeDef* handle);
+// uart_status_t uart_start(UART_HandleTypeDef* handle);
+// uart_status_t uart_stop(UART_HandleTypeDef* handle);
 uart_status_t uart_send(UART_HandleTypeDef* handle, const uint8_t* data, uint8_t length, bool blocking);
-
-
-/**
- * @brief Reads data from the user-provided RX queue
- * @param handle pointer to the UART handle
- * @param data buffer to store the received data
- * @param length of number of bytes to read
- */
 uart_status_t uart_recv(UART_HandleTypeDef* handle, uint8_t* data, uint8_t length, bool blocking);
 
 #endif /* UART_UNDEFINED */
